refactor(cli): migrate jaxa tooling onto discovery
Move jaxa's three-protocol refresher/writer/prober/emitter/validator (including the
gportal tree walker) to earthlens.jaxa.cli. Repoint the provider's refresh-walker
test at the new module; the surface tests already assert via the merged registries.
Shrink the guard allow-list (jaxa).
