+++ libs/core/src/earthlens/cli/refresh.py
@@ -748,57 +748,6 @@ def _chc_ftp_bases(catalog: Any) -> list[str]:
     )
 
 
-def _jaxa_grouped(catalog: Any) -> dict[str, list[str]]:
-    """List every live JAXA dataset id from all three protocols.
-
-    Walks the two SDKs' authoritative listings: `jaxa.earth`'s STAC
-    catalog (118 COG collections at A1 capture time) for the
-    `jaxa-earth` group, and `gportal.datasets()` (799 numeric dataset
-    ids at A1 capture time) for the `gportal` group. The `ptree` group
-    is derived **live** from the bundled JAXA catalog's `ptree` rows
-    (their `short_name` field) — P-Tree has no discoverable listing
-    endpoint, so the local catalog is the authoritative source; a
-    curator adding a new `protocol: ptree` row therefore reaches
-    `_index.yaml` automatically on the next `refresh --write`.
-
-    Args:
-        catalog: The loaded JAXA `Catalog`. Used to derive the `ptree`
-            group; the `jaxa-earth` and `gportal` groups still come
-            from the SDKs.
-
-    Returns:
-        Three-group mapping:
-        `{"jaxa-earth": [STAC collection ids], "gportal": [numeric
-        ids], "ptree": [product tokens]}`.
-    """
-    import gportal as _gportal  # type: ignore[import-not-found]
-    from jaxa.earth import je as _je  # type: ignore[import-not-found]
-
-    je_ids, _ = _je.ImageCollectionList().filter_name()
-    gp_tree = _gportal.datasets()
-
-    gp_ids: list[str] = []
-    stack: list[Any] = [gp_tree]
-    while stack:
-        node = stack.pop()
-        if isinstance(node, dict):
-            stack.extend(node.values())
-        elif isinstance(node, list):
-            gp_ids.extend(str(x) for x in node)
-    ptree_ids = sorted(
-        {
-            row.short_name
-            for row in catalog.datasets.values()
-            if row.protocol == "ptree" and row.short_name
-        }
-    )
-    return {
-        "jaxa-earth": sorted(set(str(c) for c in je_ids)),
-        "gportal": sorted(set(gp_ids)),
-        "ptree": ptree_ids,
-    }
-
-
 def _s3_grouped(catalog: Any) -> dict[str, list[str]]:
     """List the S3 registry's dataset names (its `available_datasets` universe).
 
@@ -835,7 +784,6 @@ _REFRESHERS: dict[str, Callable[[Any], dict[str, list[str]]]] = {
     "radar": _radar_grouped,
     "chc": _chc_grouped,
     "s3": _s3_grouped,
-    "jaxa": _jaxa_grouped,
 }
 
 #: Provider id -> a callable that persists a grouped live fetch back into
@@ -852,14 +800,6 @@ _WRITERS: dict[str, Callable[[BackendInfo, dict[str, list[str]]], str]] = {
     "radar": _write_radar,
     "s3": _index_writer("available_datasets"),
     "openaq": _write_openaq,
-    # JAXA's catalog YAML carries an `available_datasets:` block that lists
-    # every live id across all three protocols (jaxa-earth + gportal +
-    # ptree). The flatten=True path unions the three protocol groups
-    # into a single sorted list; the curated `datasets:` block stays
-    # hand-authored. The `ptree` group is derived from the catalog's
-    # ptree rows (see `_jaxa_grouped`) since P-Tree has no discoverable
-    # listing endpoint of its own.
-    "jaxa": _index_writer("available_datasets"),
 }
