+++ libs/core/tests/test_distribution_boundaries.py
@@ -1276,7 +1276,6 @@ class TestCliIsBackendAgnostic:
             "ecmwf",
             "gee",
             "goes",
-            "jaxa",
             "mswep",
             "nrel",
             "nwp",
